refactor: ListCollectionView, GridCollectionView 분리
